requirements.txt
mysql-connector-python~=9.0.0
python-dotenv~=1.0.1
requests~=2.32.3
tiktoken~=0.7.0
Flask~=3.0.3